Extract out common functionality into a utilities directory
